requirements.txt
nextcord~=2.6.0
mysql-connector-python~=8.1.0